==> 1.ETL.ipynb <==
#Empezamos el proceso de ETL importando las librerías necesarias

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import scipy as sp
import gzip
import io
import dask.dataframe as dd
import pyarrow as pa
import json
import re
import pyarrow.parquet as pq  # Importamos Parquet

Llamamos al primer dataset, que se encuentra en formato GNU Zip, un formato excelente para comprimir grandes cantidades de datos. Por lo tanto, es preferible usarlos sin descomprimirlos. Para ello usamos la librería GZIP. 

df = dd.read_json('Datasets\steam_games.json.gz', compression='gzip') #Usamos la librería DASK para llamar al json sin descomprimirlo
pdf = df.compute() #Utilizamos la función compute de DASK para pasar todo a un DATAFRAME de PANDAS
print(pdf.shape) #Imprimimos la cantidad de Filas y Columnas de nuestro DATAFRAME
print(pdf.head()) #Imprimimos las primeros 5 lineas de nuestro DATAFRAME

Como se puede apreciar, los datos proporcionados están llenos de nulos, por lo cual debemos empezar la limpieza.
Para ello, utilizaremos la función InfoDF, que nos proporcionará información de valor sobre el DATAFRAME.

def infoDF(pdf):
    # Cantidad de filas y columnas
    cantidad_filas = pdf.shape[0]
    cantidad_columnas = pdf.shape[1]
    
    # Cantidad de filas con datos nulos
    filas_con_nulos = pdf.isnull().any(axis=1).sum()
    
    # Porcentaje de nulos por fila y por columna
    porcentaje_nulos_filas = (pdf.isnull().sum(axis=1) / cantidad_columnas * 100).round(2)
    porcentaje_nulos_columnas = (pdf.isnull().sum() / cantidad_filas * 100).round(2)
    
    # Información general
    print("Información general del DataFrame:")
    print(f"Cantidad de filas: {cantidad_filas}")
    print(f"Cantidad de columnas: {cantidad_columnas}")
    print(f"Cantidad de filas con datos nulos: {filas_con_nulos}")
    
    # Porcentaje de nulos por fila
    print("\nPorcentaje de datos nulos por fila:")
    print(porcentaje_nulos_filas)
    
    # Porcentaje de nulos por columna
    print("\nPorcentaje de datos nulos por columna:")
    print(porcentaje_nulos_columnas)

infoDF(pdf)

Ahora, descartaremos todas las filas dentro del DATAFRAME, que contengan puros datos nulos

pdf = pdf.dropna(how='all').copy() #Borramos las filas con nulos y las reasignamos al dataframe
infoDF(pdf) #Volvemos a utilizar la función infoDF para volver a revisar los datos

Para un mejor análisis, creamos un nuevo DATAFRAME que contenga únicamente filas donde exista al menos UN (1) dato nulo

def crear_dataframe_nulos(pdf):
    # Identificar filas y columnas con datos nulos
    filas_con_nulos = pdf[pdf.isnull().any(axis=1)]
    columnas_con_nulos = pdf.loc[:, pdf.isnull().any()]
    
    # Crear un nuevo DataFrame con las filas y columnas que tienen datos nulos
    df_nulos = pd.concat([filas_con_nulos, columnas_con_nulos], axis=1)
    
    return df_nulos
df_nulos = crear_dataframe_nulos(pdf)

Luego de eliminar las filas que contenían nulos, vamos a revisar que no queden filas duplicadas

def verificar_filas_duplicadas(pdf):
    # Verificar si hay filas duplicadas
    hay_duplicados = pdf.duplicated().any()
    
    if hay_duplicados:
        print("Se encontraron filas duplicadas en el DataFrame.")
    else:
        print("No se encontraron filas duplicadas en el DataFrame.")

print(verificar_filas_duplicadas(pdf))

Parece que por ahora viene todo en orden, para continuar con la limpieza de los datos, debemos corroborar en qué tipo se encuentra cada columna. 

pdf.dtypes


Para la primera función de este proyecto, necesitamos 3 columnas cruciales: Price, Developer, Release_Date

Primero, acomodaremos la columna price, para eliminar todos sus datos nulos y corregir sus valores a tipo float

pdf = pdf.dropna(subset=['price'])
pdf.loc[:, 'price'] = pdf['price'].apply(lambda x: float(x) if pd.to_numeric(x, errors='coerce') == pd.to_numeric(x, errors='coerce') else 0)
pdf.dtypes

Una vez solucionada la columna price, pasamos a solucionar la columna release_date

Siguiendo la misma lógica, eliminamos las filas donde release_date tenga datos nulos, y analizamos si todas las fechas están en formato YYYY-MM-DD

pdf = pdf.dropna(subset=['release_date'])
# Función para identificar valores que no coinciden con el formato especificado
def encontrar_valores_incorrectos(pdf):
    try:
        pd.to_datetime(pdf['release_date'], format='%Y-%m-%d', errors='raise')
    except ValueError as e:
        valores_incorrectos = pdf.loc[pd.to_datetime(pdf['release_date'], errors='coerce').isnull(), 'release_date']
        return valores_incorrectos, str(e)
# Identificar valores que no coinciden con el formato especificado
valores_incorrectos, error_message = encontrar_valores_incorrectos(pdf)

if valores_incorrectos is not None:
    print("\nValores incorrectos en la columna 'release_date':")
    print(valores_incorrectos)
else:
    print("\nNo se encontraron valores incorrectos en la columna 'release_date'.")

if error_message:
    print("\nMensaje de error:")
    print(error_message)

ver_valores = encontrar_valores_incorrectos(pdf)

Como vemos, hay valores que no corresponden al formato YYYY-MM-DD y los debemos tratar por separado

# Función para extraer el año
def extraer_ano(fecha):
    # Intentamos extraer el año del formato 'YYYY-MM-DD'
    try:
        return int(fecha.split('-')[0])
    except:
        pass
    
    # Intentamos extraer el año del formato 'MMM YYYY'
    try:
        return int(fecha.split()[-1])
    except:
        pass
    
    # Si no podemos extraer el año, devolvemos 0
    return 0

# Aplicamos la función a la columna 'release_date'
pdf['release_date'] = pdf['release_date'].apply(extraer_ano)

Ahora, como podemos ver a continuación, la columna price y la columna release_date se encuentran en formato numérico, para un mejor manejo

pdf.dtypes

Para continuar el proceso de limpieza, debemos tratar de unificar las columnas genres y tags, ya que, como podemos ver a continuación, las mismas los datos son muy similares y se complementan.

pdf[['genres', 'tags']]

# Reemplazar los valores nulos en la columna 'genres' con una cadena vacía
pdf['genres'].fillna('', inplace=True)

# Reemplazar los valores nulos en la columna 'tags' con una cadena vacía
pdf['tags'].fillna('', inplace=True)

# Función para combinar los valores de 'genres' y 'tags'
def combine_genres_tags(row):
    genres = row['genres']
    tags = row['tags']
    
    # Si 'genres' está vacío, asigna los valores de 'tags'
    if not genres:
        return tags
    
    # Si 'tags' tiene más información que 'genres', combina ambas
    if tags:
        tags_list = tags.split(', ')
        for tag in tags_list:
            if tag not in genres:
                genres += ', ' + tag
    return genres

# Aplicar la función a cada fila del DataFrame
pdf['genres'] = pdf.apply(combine_genres_tags, axis=1)


Con esto, ya podemos solucionar el primer ENDPOINT solicitado, para el cual necesitábamos ingresar un string con un Developer, y nos arrojará el porcentaje de juegos gratuitos que salieron cada año


table = pa.Table.from_pandas(pdf)
# Escribimos las tabla en un archivo Parquet
pq.write_table(table, 'Datasets/pdf_SteamGames.parquet')


Vamos con el segundo DATAFRAME, en este caso, Steam Games. El mismo, tiene errores de escritura por lo cual debemos cambiar el proceso de carga de datos.

import ast

#Creamos una lista vacía llamada "rows" donde almacenaremos los datos del archivo JSON.
rows = []
#Abrir el archivo "user_reviews.json/australian_user_reviews.json" con la codificación MacRoman.
with open(r"Datasets\user_reviews.json", encoding='utf-8') as f:
    # Leer cada línea del archivo.
    for line in f.readlines():
        # Utilizar "ast.literal_eval" para convertir cada línea en un diccionario de Python
        # y agregarlo a la lista "rows".
        rows.append(ast.literal_eval(line))

#Crear un DataFrame de Pandas a partir de la lista de diccionarios "rows".
df_user_reviews = pd.DataFrame(rows)
#Veamos unos registros al asar
df_user_reviews.sample(2)

Repetimos función para analizar el estado del Dataframe

infoDF(df_user_reviews)

Este dataframe no contiene datos nulos pero nos encontramos con un nuevo desafío. La fila de reviews es una cantidad de listas anidadas que debemos desanidar

print("Columna reviews:")
print(df_user_reviews['reviews'].iloc[0])  # Imprimir un valor específico de la columna 'reviews'
print(df_user_reviews['reviews'].apply(type).unique())  # Verificar el tipo de dato de toda la columna 'reviews'

Para que nuestro Dataframe tenga sentido, debemos desarmar esta lista y crear columnas por separado

# Crear una lista vacía para almacenar las nuevas filas
new_rows = []

# Iterar sobre cada fila del DataFrame original
for index, row in df_user_reviews.iterrows():
    # Obtener la lista de reviews de la fila actual
    reviews_list = row.get('reviews')
    # Verificar si la lista de reviews está presente y no es None
    if reviews_list is not None:
        # Iterar sobre cada review en la lista
        for review_dict in reviews_list:
            # Crear una nueva fila con los datos de la review y otros datos de la fila original
            new_row = {
                'user_id': row['user_id'],
                'user_url': row['user_url'],
                'funny': review_dict.get('funny'),
                'posted': review_dict.get('posted'),
                'last_edited': review_dict.get('last_edited'),
                'item_id': review_dict.get('item_id'),
                'helpful': review_dict.get('helpful'),
                'recommend': review_dict.get('recommend'),
                'review': review_dict.get('review')
            }
            # Agregar la nueva fila a la lista de filas
            new_rows.append(new_row)

# Crear un nuevo DataFrame con las nuevas filas
new_df_user_reviews = pd.DataFrame(new_rows)

# Ver el nuevo DataFrame
print(new_df_user_reviews)

Mucho mejor, ahora corroboramos que no existan filas duplicadas

filas_duplicadas = new_df_user_reviews[new_df_user_reviews.duplicated()]
num_filas_duplicadas = len(filas_duplicadas)
if num_filas_duplicadas > 0:
    print("Se encontraron", num_filas_duplicadas, "filas duplicadas en el DataFrame.")
else:
    print("No se encontraron filas duplicadas en el DataFrame.")

Se encontraron muchas, asi que toca eliminarlas.

# Eliminar filas duplicadas en todo el DataFrame
new_df_user_reviews_sin_duplicados = new_df_user_reviews.drop_duplicates()

Todo listo, ahora eliminamos las columnas que no tienen valor para nuestro futuro análisis.

# Eliminar columnas usando el método drop()
columnas_a_eliminar = ['funny', 'last_edited', 'helpful']
new_df_user_reviews_sin_duplicados.drop(columnas_a_eliminar, axis=1, inplace=True)

# Verificar que las columnas se han eliminado
print(new_df_user_reviews_sin_duplicados.columns)

Y empezamos la corrección de tipos de datos


print(new_df_user_reviews_sin_duplicados.dtypes)

Lo primero que debemos corregir es la columna 'posted', ya que la misma no coincide dentro de los datos y nos complicaría el funcionamiento de la función. Para ello crearemos una nueva columna 'year' donde extraeremos únicamente el año en que fue publicada la reseña.

# Función para extraer el año de la columna 'posted'
def extract_year(date_string):
    # Expresión regular para encontrar el año
    year_pattern = r'\b\d{4}\b'
    # Buscar el año en el texto
    year_match = re.search(year_pattern, date_string)
    if year_match:
        return int(year_match.group())
    else:
        return None

# Aplicar la función para extraer el año y crear una nueva columna 'year'
new_df_user_reviews_sin_duplicados['year'] = new_df_user_reviews_sin_duplicados['posted'].apply(extract_year)

# Calcular el promedio de los años presentes en el dataframe
average_year = new_df_user_reviews_sin_duplicados['year'].mean()

# Rellenar los valores de año faltantes con el promedio
new_df_user_reviews_sin_duplicados['year'].fillna(average_year, inplace=True)

# Convertir los años a enteros
new_df_user_reviews_sin_duplicados['year'] = new_df_user_reviews_sin_duplicados['year'].astype(int)

# Imprimir el dataframe resultante
print(new_df_user_reviews_sin_duplicados)

new_df_user_reviews_sin_duplicados.dtypes
new_df_user_reviews_sin_duplicados.drop(columns=['posted'], inplace=True)

# Convertir la columna 'item_id' a tipo float
new_df_user_reviews_sin_duplicados['item_id'] = new_df_user_reviews_sin_duplicados['item_id'].astype(float)

new_df_user_reviews_sin_duplicados.dtypes

table = pa.Table.from_pandas(new_df_user_reviews_sin_duplicados)
# Escribimos las tabla en un archivo Parquet
pq.write_table(table, 'Datasets/new_user_reviews.parquet')

Guardamos la nueva tabla en parquet para su futuro uso

Empezamos el último ETL, del json user_items

import gzip

# Lista para almacenar los datos del archivo JSON
rows = []

# Abrir el archivo comprimido en modo lectura de bytes ('rb')
with gzip.open(r"Datasets\users_items.json.gz", 'rb') as f:
    # Decodificar cada línea del archivo como texto UTF-8 y agregarla a la lista 'rows'
    for line in f:
        rows.append(line.decode('utf-8'))

# Crear un DataFrame de Pandas a partir de la lista de strings
df_user_items = pd.DataFrame(rows)

Corroboramos la información de nuestro nuevo Dataframe

df_user_items.head()
infoDF(df_user_items)

Como podemos ver, aquí el error se encuentra en que existe únicamente una sola columna con toda la información anidada dentro.

print(df_user_items.columns)

Comenzamos el proceso de desanidar todo nuestro dataframe para poder utilizarlo

# Crear una función para procesar cada fila del DataFrame original
def process_row(row):
    # Convertir el string de la columna 0 a un diccionario
    df_user_items = ast.literal_eval(row.iloc[0])
    
    # Extraer información
    user_id = df_user_items['user_id']
    items_count = df_user_items['items_count']
    steam_id = df_user_items['steam_id']
    user_url = df_user_items['user_url']
    
    # Crear una lista de diccionarios para los 'items'
    items_list = df_user_items['items']
    
    # Crear una lista para almacenar cada fila de datos
    rows = []
    
    # Iterar sobre los elementos de la lista de items
    for item in items_list:
        # Crear un diccionario con la información que necesitas
        row_df_user_items = {
            'user_id': user_id,
            'items_count': items_count,
            'steam_id': steam_id,
            'user_url': user_url,
            'item_id': item.get('item_id', None),  # Manejar casos donde el valor no está presente
            'item_name': item.get('item_name', None),
            'playtime_forever': item.get('playtime_forever', None),
            'playtime_2weeks': item.get('playtime_2weeks', None)
        }
        # Agregar el diccionario a la lista de filas
        rows.append(row_df_user_items)
    
    # Retornar la lista de filas
    return rows

# Aplicar la función a cada fila del DataFrame original y obtener una lista de listas de diccionarios
processed_data = df_user_items.apply(process_row, axis=1).tolist()

# Convertir la lista de listas de diccionarios en un solo DataFrame
new_df = pd.DataFrame([item for sublist in processed_data for item in sublist])

# Mostrar el nuevo DataFrame
print(new_df)

Ahora, podemos empezar a trabajar sobre nuestro dataframe

filas_duplicadas = new_df[new_df.duplicated()]
num_filas_duplicadas = len(filas_duplicadas)
if num_filas_duplicadas > 0:
    print("Se encontraron", num_filas_duplicadas, "filas duplicadas en el DataFrame.")
else:
    print("No se encontraron filas duplicadas en el DataFrame.")

Eliminamos los duplicados y vemos el tipo de datos que contiene

new_df_sin_duplicados = new_df.drop_duplicates()

new_df_sin_duplicados.dtypes

Para trabajar nuestros dataframes en conjunto, necesitamos que item_id sea tipo float y que item_name sea tipo string

# Convertir la columna 'item_id' a tipo float
new_df_sin_duplicados['item_id'] = new_df_sin_duplicados['item_id'].astype(float)

# Convertir la columna 'item_name' a tipo string
new_df_sin_duplicados['item_name'] = new_df_sin_duplicados['item_name'].astype(str)

Todo listo, podemos guardar nuestros dataframes en Parquet, y continuar el proyecto

table = pa.Table.from_pandas(new_df_sin_duplicados)
# Escribimos las tabla en un archivo Parquet
pq.write_table(table, 'Datasets/new_df_users_items.parquet')
